Move to pkg/duck/v1beta1.
