+++ pkg/apis/events/v1beta1/register.go
@@ -53,8 +53,6 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&CloudSchedulerSourceList{},
 		&CloudPubSubSource{},
 		&CloudPubSubSourceList{},
-		&CloudBuildSource{},
-		&CloudBuildSourceList{},
 	)
 	metav1.AddToGroupVersion(scheme, SchemeGroupVersion)
 	return nil
