+++ pkg/duck/v1beta1/identifiable.go
@@ -14,21 +14,21 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1alpha1
+package v1beta1
 
 import (
 	"knative.dev/pkg/apis"
 	"knative.dev/pkg/kmeta"
 
-	duckv1alpha1 "github.com/google/knative-gcp/pkg/apis/duck/v1alpha1"
+	duckv1beta1 "github.com/google/knative-gcp/pkg/apis/duck/v1beta1"
 )
 
 type Identifiable interface {
 	kmeta.OwnerRefable
 	// IdentitySpec returns the IdentitySpec portion of the Spec.
-	IdentitySpec() *duckv1alpha1.IdentitySpec
+	IdentitySpec() *duckv1beta1.IdentitySpec
 	// IdentityStatus returns the IdentityStatus portion of the Status.
-	IdentityStatus() *duckv1alpha1.IdentityStatus
+	IdentityStatus() *duckv1beta1.IdentityStatus
 	// ConditionSet returns the apis.ConditionSet of the embedding object
 	ConditionSet() *apis.ConditionSet
 }
